refactor: domain과 view 역할 분리
