--- src/main/kotlin/blackjack/view/GameView.kt
@@ -1,9 +1,0 @@
-package blackjack.view
-
-import blackjack.domain.game.Game
-import blackjack.util.inputPlayers
-
-fun main() {
-    val game = Game(inputPlayers())
-    game.start()
-}
