+++ src/main/kotlin/blackjack/view/ResultView.kt
@@ -1,0 +1,10 @@
+package blackjack.view
+
+import blackjack.domain.player.Player
+
+object ResultView {
+    fun printResult(players: List<Player>) {
+        println()
+        players.forEach { player -> println("$player - 결과: ${player.score()}") }
+    }
+}
